refactor: extract store flags
